move classes to seperate directory
