+++ potato/classes/active_learning_state.py
@@ -1,0 +1,16 @@
+class ActiveLearningState:
+    """
+    A class for maintaining state on active learning.
+    """
+
+    def __init__(self):
+        self.id_to_selection_type = {}
+        self.id_to_update_round = {}
+        self.cur_round = 0
+
+    def update_selection_types(self, id_to_selection_type):
+        self.cur_round += 1
+
+        for iid, st in id_to_selection_type.items():
+            self.id_to_selection_type[iid] = st
+            self.id_to_update_round[iid] = self.cur_round
